+++ template/client/src/main.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+import ReactDOM from "react-dom/client";
+
+import { RouterProvider, createBrowserRouter } from "react-router-dom";
+
+import App from "./App";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <App />,
+  },
+]);
+
+const rootElement = document.getElementById("root");
+
+if (rootElement != null) {
+  const root = ReactDOM.createRoot(rootElement);
+
+  root.render(
+    <React.StrictMode>
+      <RouterProvider router={router} />
+    </React.StrictMode>,
+  );
+}